restaurant-reviews/pyproject.toml
[project]
name = "restaurant-reviews"
version = "0.1.0"
description = ""
authors = [
    {name = "JaeheeLee-be",email = "clodusky1@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "mysqlclient (>=2.2.8,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
